Move license lightbox outside the card.
