+++ client/jetpack-cloud/sections/partner-portal/license-bundle-card/index.tsx
@@ -1,10 +1,13 @@
 import { Button } from '@automattic/components';
 import formatCurrency from '@automattic/format-currency';
 import { useTranslate } from 'i18n-calypso';
-import { useCallback } from 'react';
+import { useCallback, useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { APIProductFamilyProduct } from '../../../../state/partner-portal/types';
 import { useProductDescription } from '../hooks';
-import LicenseLightbox from '../license-lightbox-wrapper';
+import LicenseLightbox from '../license-lightbox';
+import LicenseLightboxLink from '../license-lightbox-link';
 import { getProductTitle } from '../utils';
 
 import './style.scss';
@@ -18,7 +21,9 @@ interface Props {
 export default function LicenseBundleCard( props: Props ) {
 	const { tabIndex, product, onSelectProduct } = props;
 	const productTitle = getProductTitle( product.name );
+	const [ showLightbox, setShowLightbox ] = useState( false );
 	const translate = useTranslate();
+	const dispatch = useDispatch();
 
 	const onSelect = useCallback( () => {
 		onSelectProduct( product );
@@ -26,41 +31,67 @@ export default function LicenseBundleCard( props: Props ) {
 
 	const { description: productDescription } = useProductDescription( product.slug );
 
+	const onShowLightbox = useCallback(
+		( e: React.MouseEvent< HTMLElement > ) => {
+			e.stopPropagation();
+
+			dispatch(
+				recordTracksEvent( 'calypso_partner_portal_issue_license_product_view', {
+					product: product.slug,
+				} )
+			);
+
+			setShowLightbox( true );
+		},
+		[ dispatch, product ]
+	);
+
+	const onHideLightbox = useCallback( () => {
+		setShowLightbox( false );
+	}, [] );
+
 	return (
-		<div className="license-bundle-card">
-			<div className="license-bundle-card__details">
-				<h3 className="license-bundle-card__title">{ productTitle }</h3>
+		<>
+			<div className="license-bundle-card">
+				<div className="license-bundle-card__details">
+					<h3 className="license-bundle-card__title">{ productTitle }</h3>
 
-				<div className="license-bundle-card__description">{ productDescription }</div>
+					<div className="license-bundle-card__description">{ productDescription }</div>
 
+					<LicenseLightboxLink productName={ productTitle } onClick={ onShowLightbox } />
+				</div>
+
+				<div className="license-bundle-card__footer">
+					<div className="license-bundle-card__pricing">
+						<div className="license-bundle-card__price">
+							{ formatCurrency( product.amount, product.currency ) }
+						</div>
+						<div className="license-bundle-card__price-interval">
+							{ product.price_interval === 'day' && translate( '/USD per license per day' ) }
+							{ product.price_interval === 'month' && translate( '/USD per license per month' ) }
+						</div>
+					</div>
+
+					<Button
+						primary
+						className="license-bundle-card__select-license"
+						onClick={ onSelect }
+						tabIndex={ tabIndex }
+					>
+						{ translate( 'Select' ) }
+					</Button>
+				</div>
+			</div>
+
+			{ showLightbox && (
 				<LicenseLightbox
 					product={ product }
 					ctaLabel={ translate( 'Select License' ) }
 					onActivate={ onSelect }
+					onClose={ onHideLightbox }
 				/>
-			</div>
-
-			<div className="license-bundle-card__footer">
-				<div className="license-bundle-card__pricing">
-					<div className="license-bundle-card__price">
-						{ formatCurrency( product.amount, product.currency ) }
-					</div>
-					<div className="license-bundle-card__price-interval">
-						{ product.price_interval === 'day' && translate( '/USD per license per day' ) }
-						{ product.price_interval === 'month' && translate( '/USD per license per month' ) }
-					</div>
-				</div>
-
-				<Button
-					primary
-					className="license-bundle-card__select-license"
-					onClick={ onSelect }
-					tabIndex={ tabIndex }
-				>
-					{ translate( 'Select' ) }
-				</Button>
-			</div>
-		</div>
+			) }
+		</>
 	);
 }
