+++ client/jetpack-cloud/sections/partner-portal/license-product-card/index.tsx
@@ -2,10 +2,13 @@ import { Gridicon } from '@automattic/components';
 import formatCurrency from '@automattic/format-currency';
 import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
-import { useCallback, useEffect } from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { APIProductFamilyProduct } from '../../../../state/partner-portal/types';
 import { useProductDescription } from '../hooks';
-import LicenseLightbox from '../license-lightbox-wrapper';
+import LicenseLightbox from '../license-lightbox';
+import LicenseLightboxLink from '../license-lightbox-link';
 import { getProductTitle } from '../utils';
 
 import './style.scss';
@@ -31,7 +34,9 @@ export default function LicenseProductCard( props: Props ) {
 		isMultiSelect,
 	} = props;
 	const productTitle = getProductTitle( product.name );
+	const [ showLightbox, setShowLightbox ] = useState( false );
 	const translate = useTranslate();
+	const dispatch = useDispatch();
 
 	const onSelect = useCallback( () => {
 		if ( isDisabled ) {
@@ -64,60 +69,84 @@ export default function LicenseProductCard( props: Props ) {
 
 	const { description: productDescription } = useProductDescription( product.slug );
 
+	const onShowLightbox = useCallback(
+		( e: React.MouseEvent< HTMLElement > ) => {
+			e.stopPropagation();
+
+			dispatch(
+				recordTracksEvent( 'calypso_partner_portal_issue_license_product_view', {
+					product: product.slug,
+				} )
+			);
+
+			setShowLightbox( true );
+		},
+		[ dispatch, product ]
+	);
+
+	const onHideLightbox = useCallback( () => {
+		setShowLightbox( false );
+	}, [] );
+
 	return (
-		<div
-			onClick={ onSelect }
-			onKeyDown={ onKeyDown }
-			role={ isMultiSelect ? 'checkbox' : 'radio' }
-			tabIndex={ tabIndex }
-			aria-checked={ isSelected }
-			aria-disabled={ isDisabled }
-			className={ classNames( {
-				'license-product-card': true,
-				selected: isSelected,
-				disabled: isDisabled,
-			} ) }
-		>
-			<div className="license-product-card__inner">
-				<div className="license-product-card__details">
-					<div className="license-product-card__main">
-						<div className="license-product-card__heading">
-							<h3 className="license-product-card__title">{ productTitle }</h3>
+		<>
+			<div
+				onClick={ onSelect }
+				onKeyDown={ onKeyDown }
+				role={ isMultiSelect ? 'checkbox' : 'radio' }
+				tabIndex={ tabIndex }
+				aria-checked={ isSelected }
+				aria-disabled={ isDisabled }
+				className={ classNames( {
+					'license-product-card': true,
+					selected: isSelected,
+					disabled: isDisabled,
+				} ) }
+			>
+				<div className="license-product-card__inner">
+					<div className="license-product-card__details">
+						<div className="license-product-card__main">
+							<div className="license-product-card__heading">
+								<h3 className="license-product-card__title">{ productTitle }</h3>
 
-							<div className="license-product-card__description">{ productDescription }</div>
+								<div className="license-product-card__description">{ productDescription }</div>
 
-							<LicenseLightbox
-								product={ product }
-								ctaLabel={
-									isSelected ? translate( 'Unselect License' ) : translate( 'Select License' )
-								}
-								isCTAPrimary={ ! isSelected }
-								isDisabled={ isDisabled }
-								onActivate={ onSelectProduct }
-							/>
+								<LicenseLightboxLink productName={ productTitle } onClick={ onShowLightbox } />
+							</div>
+
+							<div
+								className={ classNames( 'license-product-card__select-button', {
+									'license-product-card_multi-select': isMultiSelect,
+								} ) }
+							>
+								{ isSelected && <Gridicon icon="checkmark" /> }
+							</div>
 						</div>
 
-						<div
-							className={ classNames( 'license-product-card__select-button', {
-								'license-product-card_multi-select': isMultiSelect,
-							} ) }
-						>
-							{ isSelected && <Gridicon icon="checkmark" /> }
-						</div>
-					</div>
-
-					<div className="license-product-card__pricing">
-						<div className="license-product-card__price">
-							{ formatCurrency( product.amount, product.currency ) }
-						</div>
-						<div className="license-product-card__price-interval">
-							{ product.price_interval === 'day' && translate( '/USD per license per day' ) }
-							{ product.price_interval === 'month' && translate( '/USD per license per month' ) }
+						<div className="license-product-card__pricing">
+							<div className="license-product-card__price">
+								{ formatCurrency( product.amount, product.currency ) }
+							</div>
+							<div className="license-product-card__price-interval">
+								{ product.price_interval === 'day' && translate( '/USD per license per day' ) }
+								{ product.price_interval === 'month' && translate( '/USD per license per month' ) }
+							</div>
 						</div>
 					</div>
 				</div>
 			</div>
-		</div>
+
+			{ showLightbox && (
+				<LicenseLightbox
+					product={ product }
+					ctaLabel={ isSelected ? translate( 'Unselect License' ) : translate( 'Select License' ) }
+					isCTAPrimary={ ! isSelected }
+					isDisabled={ isDisabled }
+					onActivate={ onSelectProduct }
+					onClose={ onHideLightbox }
+				/>
+			) }
+		</>
 	);
 }
